django-server/requirements.txt
asgiref==3.7.2
Django==3.2.7
pip==23.3.1
setuptools==65.5.0
sqlparse==0.4.4
mysqlclient==2.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1